ssz: simple sszSerializer builder introduced
