+++ types/src/main/java/org/ethereum/beacon/types/ssz/Serializer.java
@@ -1,37 +1,33 @@
 package org.ethereum.beacon.types.ssz;
 
-import org.ethereum.beacon.util.ssz.SSZAnnotationSchemeBuilder;
-import org.ethereum.beacon.util.ssz.SSZSchemeBuilder;
 import org.ethereum.beacon.util.ssz.SSZSerializer;
-import org.ethereum.beacon.util.ssz.type.SSZEncoderDecoder;
+import org.ethereum.beacon.util.ssz.SSZSerializerBuilder;
 import tech.pegasys.artemis.util.bytes.BytesValue;
 
-public class Serializer extends SSZSerializer {
+public class Serializer {
 
-  private final static Serializer ANNOTATION_SERIALIZER = new Serializer(new SSZAnnotationSchemeBuilder())
-      .withSSZEncoderDecoder(new SSZHash())
-      .withSSZEncoderDecoder(new SSZUInt())
-      .withSSZEncoderDecoder(new SSZBytesValue());
+  private final static SSZSerializer ANNOTATION_SERIALIZER;
+  static {
+    SSZSerializerBuilder builder = SSZSerializerBuilder.getBakedAnnotationBuilder();
+    builder.addEncoderDecoder(new SSZHash());
+    builder.addEncoderDecoder(new SSZUInt());
+    builder.addEncoderDecoder(new SSZBytesValue());
+    ANNOTATION_SERIALIZER = builder.build();
+  }
+  private final static Serializer INSTANCE = new Serializer();
 
-  public Serializer(SSZSchemeBuilder schemeBuilder) {
-    super(schemeBuilder);
+  public Serializer() {
   }
 
   public static Serializer annotationSerializer() {
-    return ANNOTATION_SERIALIZER;
-  }
-
-  @Override
-  public Serializer withSSZEncoderDecoder(SSZEncoderDecoder encoderDecoder) {
-    super.withSSZEncoderDecoder(encoderDecoder);
-    return this;
+    return INSTANCE;
   }
 
   public BytesValue encode2(Object input) {
-    return BytesValue.wrap(encode(input));
+    return BytesValue.wrap(ANNOTATION_SERIALIZER.encode(input));
   }
 
   public Object decode(BytesValue data, Class clazz) {
-    return decode(data.getArrayUnsafe(), clazz);
+    return ANNOTATION_SERIALIZER.decode(data.getArrayUnsafe(), clazz);
   }
 }
